+++ ibis/impala/compiler.py
@@ -4,12 +4,12 @@ from typing import Optional
 
 import ibis
 import ibis.backends.base_sqlalchemy.compiler as comp
+import ibis.backends.base_sqlalchemy.transforms as transforms
 import ibis.common.exceptions as com
 import ibis.expr.analysis as L
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
 import ibis.expr.types as ir
-import ibis.sql.transforms as transforms
 import ibis.util as util
 from ibis.backends import base_sql
 from ibis.backends.base_sql import (
